update-4.2

+++ Program.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using System.Numerics;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -18,7 +19,7 @@ public class LCMController : ControllerBase
     [HttpGet]
     public IActionResult GetLCM([FromQuery] string x = "", [FromQuery] string y = "")
     {
-        if (!IsNaturalNumber(x, out long xVal) || !IsNaturalNumber(y, out long yVal))
+        if (!IsNaturalNumber(x, out double xVal) || !IsNaturalNumber(y, out double yVal))
         {
             return Content("NaN", "text/plain");
         }
@@ -27,9 +28,9 @@ public class LCMController : ControllerBase
         return Content(lcm.ToString(), "text/plain");
     }
 
-    private static bool IsNaturalNumber(string input, out long value)
+    private static bool IsNaturalNumber(string input, out double value)
     {
-        if (long.TryParse(input, out value) && value > 0)
+        if (double.TryParse(input, out value) && value > 0)
         {
             return true;
         }
